refactor: split repo cell read payloads
